Initial commit of slict: a sliceable dictionary.
This defines the interfaces and provides a working but low-performance
implementation of it.

sd = Slict(d) stores a handle to a dictionary d and provides a sliceable
read-only interface to it.

sd1 = sd[foo, :] returns a new slict that stores the first index "key" and
appends it to keys before passing them to d.  That is, sd1[bar] passes
(foo, bar) to the original dictionary d.